#21 - moved box classes into new package
