+++ src/main/java/org/opengis/cite/gmljpx20/core/CoreTests.java
@@ -25,14 +25,14 @@ import org.opengis.cite.gmljpx20.GMLJP2;
 import org.opengis.cite.gmljpx20.SuiteAttribute;
 import org.opengis.cite.gmljpx20.util.TestSuiteLogger;
 import org.opengis.cite.gmljpx20.util.XMLUtils;
-import org.opengis.cite.gmljpx20.util.jp2.Association;
-import org.opengis.cite.gmljpx20.util.jp2.Box;
-import org.opengis.cite.gmljpx20.util.jp2.ContigousCodestream;
-import org.opengis.cite.gmljpx20.util.jp2.FileType;
+import org.opengis.cite.gmljpx20.box.Association;
+import org.opengis.cite.gmljpx20.box.Box;
+import org.opengis.cite.gmljpx20.box.ContigousCodestream;
+import org.opengis.cite.gmljpx20.box.FileType;
 import org.opengis.cite.gmljpx20.util.jp2.JP2Stream;
-import org.opengis.cite.gmljpx20.util.jp2.Label;
-import org.opengis.cite.gmljpx20.util.jp2.ResourceRequirements;
-import org.opengis.cite.gmljpx20.util.jp2.XMLBox;
+import org.opengis.cite.gmljpx20.box.Label;
+import org.opengis.cite.gmljpx20.box.ResourceRequirements;
+import org.opengis.cite.gmljpx20.box.XMLBox;
 import org.testng.Assert;
 import org.testng.ITestContext;
 import org.testng.SkipException;
